refactor(model): extract CUDA Graph capture/replay into shared CudaGraphState
Model decode logic no longer knows about CUDA Graph — it only provides
the kernel closure. Eliminates duplicated capture/replay code across
Qwen3 and Qwen3.5.
