+++ src/model.rs
@@ -6,6 +6,7 @@ use rand::rngs::StdRng;
 use crate::sampler::SamplingParams;
 use crate::tensor::DeviceVec;
 
+pub(crate) mod cuda_graph;
 mod kv_cache;
 
 pub mod qwen3;
